(1248) Split up userActivityModel
